style: update editor
